requirements.txt
pillow==11.1.0
opencv-python==4.11.0.86
numpy==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
pytesseract==0.3.13
scipy==1.15.2
imageio==2.37.0
customtkinter==5.2.2
tqdm
colorlog